split the module into four parts
